app/requirements.txt
asgiref==3.8.1
Django==5.1
django-elasticsearch-dsl==8.0
djangorestframework==3.15.2
gunicorn==23.0.0
packaging==24.1
psycopg2-binary==2.9.9
sqlparse==0.5.1
isodate==0.6.1
pyparsing==3.1.2
rdflib==7.0.0
polars==1.5.0
